Create HDF5 package extension
