+++ src/ITensorMPS/abstractmps.jl
@@ -699,6 +699,26 @@ function map(f::Function, M::AbstractMPS; set_limits::Bool=true)
   return map!(f, copy(M); set_limits=set_limits)
 end
 
+import ITensors:
+  dag,
+  #dag!, 
+  prime,
+  prime!,
+  setprime,
+  setprime!,
+  noprime,
+  noprime!,
+  swapprime,
+  swapprime!,
+  replaceprime,
+  replaceprime!,
+  addtags,
+  addtags!,
+  removetags,
+  removetags!,
+  settags,
+  settags!
+
 for fname in (
   :dag,
   :prime,
