+++ src/ITensorMPS/mps.jl
@@ -1,5 +1,4 @@
 using Adapt: adapt
-using HDF5: attributes, create_group, open_group
 using NDTensors: using_auto_fermion
 using Random: Random
 
@@ -1031,29 +1030,4 @@ end
 
 function expect(psi::MPS, op1::Matrix{<:Number}, ops::Matrix{<:Number}...; kwargs...)
   return expect(psi, (op1, ops...); kwargs...)
-end
-
-function HDF5.write(parent::Union{HDF5.File,HDF5.Group}, name::AbstractString, M::MPS)
-  g = create_group(parent, name)
-  attributes(g)["type"] = "MPS"
-  attributes(g)["version"] = 1
-  N = length(M)
-  write(g, "length", N)
-  write(g, "rlim", M.rlim)
-  write(g, "llim", M.llim)
-  for n in 1:N
-    write(g, "MPS[$(n)]", M[n])
-  end
-end
-
-function HDF5.read(parent::Union{HDF5.File,HDF5.Group}, name::AbstractString, ::Type{MPS})
-  g = open_group(parent, name)
-  if read(attributes(g)["type"]) != "MPS"
-    error("HDF5 group or file does not contain MPS data")
-  end
-  N = read(g, "length")
-  rlim = read(g, "rlim")
-  llim = read(g, "llim")
-  v = [read(g, "MPS[$(i)]", ITensor) for i in 1:N]
-  return MPS(v, llim, rlim)
 end
